Move HoneyPoolInterface and HoneyPool to models/game/pools
