+++ frontend/models/game/base/DoctorInterface.php
@@ -9,6 +9,8 @@
 namespace frontend\models\game\base;
 
 
+use frontend\models\game\pools\interfaces\HoneyPoolInterface;
+
 interface DoctorInterface
 {
     public function heal(BeeInterface $bee, HoneyPoolInterface $pool);
