+++ frontend/models/game/characters/Worker.php
@@ -10,7 +10,7 @@ namespace frontend\models\game\characters;
 
 
 use frontend\models\game\base\Bee;
-use frontend\models\game\base\HoneyPoolInterface;
+use frontend\models\game\pools\interfaces\HoneyPoolInterface;
 use frontend\models\game\characters\interfaces\WorkerInterface;
 
 class Worker extends Bee implements WorkerInterface
