+++ tests/fixtures/GameWithoutCharacters.php
@@ -9,7 +9,7 @@
 namespace tests\fixtures;
 
 
-use frontend\models\game\base\HoneyPool;
+use frontend\models\game\pools\HoneyPool;
 use frontend\models\game\pools\CharacterPool;
 use frontend\models\game\Game;
 use frontend\models\game\GameInterface;
